Module aws_sdk_devicefarm::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateDevicePool

Fluent builder constructing a request to CreateDevicePool.

CreateInstanceProfile

Fluent builder constructing a request to CreateInstanceProfile.

CreateNetworkProfile

Fluent builder constructing a request to CreateNetworkProfile.

CreateProject

Fluent builder constructing a request to CreateProject.

CreateRemoteAccessSession

Fluent builder constructing a request to CreateRemoteAccessSession.

CreateTestGridProject

Fluent builder constructing a request to CreateTestGridProject.

CreateTestGridUrl

Fluent builder constructing a request to CreateTestGridUrl.

CreateUpload

Fluent builder constructing a request to CreateUpload.

CreateVPCEConfiguration

Fluent builder constructing a request to CreateVPCEConfiguration.

DeleteDevicePool

Fluent builder constructing a request to DeleteDevicePool.

DeleteInstanceProfile

Fluent builder constructing a request to DeleteInstanceProfile.

DeleteNetworkProfile

Fluent builder constructing a request to DeleteNetworkProfile.

DeleteProject

Fluent builder constructing a request to DeleteProject.

DeleteRemoteAccessSession

Fluent builder constructing a request to DeleteRemoteAccessSession.

DeleteRun

Fluent builder constructing a request to DeleteRun.

DeleteTestGridProject

Fluent builder constructing a request to DeleteTestGridProject.

DeleteUpload

Fluent builder constructing a request to DeleteUpload.

DeleteVPCEConfiguration

Fluent builder constructing a request to DeleteVPCEConfiguration.

GetAccountSettings

Fluent builder constructing a request to GetAccountSettings.

GetDevice

Fluent builder constructing a request to GetDevice.

GetDeviceInstance

Fluent builder constructing a request to GetDeviceInstance.

GetDevicePool

Fluent builder constructing a request to GetDevicePool.

GetDevicePoolCompatibility

Fluent builder constructing a request to GetDevicePoolCompatibility.

GetInstanceProfile

Fluent builder constructing a request to GetInstanceProfile.

GetJob

Fluent builder constructing a request to GetJob.

GetNetworkProfile

Fluent builder constructing a request to GetNetworkProfile.

GetOfferingStatus

Fluent builder constructing a request to GetOfferingStatus.

GetProject

Fluent builder constructing a request to GetProject.

GetRemoteAccessSession

Fluent builder constructing a request to GetRemoteAccessSession.

GetRun

Fluent builder constructing a request to GetRun.

GetSuite

Fluent builder constructing a request to GetSuite.

GetTest

Fluent builder constructing a request to GetTest.

GetTestGridProject

Fluent builder constructing a request to GetTestGridProject.

GetTestGridSession

Fluent builder constructing a request to GetTestGridSession.

GetUpload

Fluent builder constructing a request to GetUpload.

GetVPCEConfiguration

Fluent builder constructing a request to GetVPCEConfiguration.

InstallToRemoteAccessSession

Fluent builder constructing a request to InstallToRemoteAccessSession.

ListArtifacts

Fluent builder constructing a request to ListArtifacts.

ListDeviceInstances

Fluent builder constructing a request to ListDeviceInstances.

ListDevicePools

Fluent builder constructing a request to ListDevicePools.

ListDevices

Fluent builder constructing a request to ListDevices.

ListInstanceProfiles

Fluent builder constructing a request to ListInstanceProfiles.

ListJobs

Fluent builder constructing a request to ListJobs.

ListNetworkProfiles

Fluent builder constructing a request to ListNetworkProfiles.

ListOfferingPromotions

Fluent builder constructing a request to ListOfferingPromotions.

ListOfferingTransactions

Fluent builder constructing a request to ListOfferingTransactions.

ListOfferings

Fluent builder constructing a request to ListOfferings.

ListProjects

Fluent builder constructing a request to ListProjects.

ListRemoteAccessSessions

Fluent builder constructing a request to ListRemoteAccessSessions.

ListRuns

Fluent builder constructing a request to ListRuns.

ListSamples

Fluent builder constructing a request to ListSamples.

ListSuites

Fluent builder constructing a request to ListSuites.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ListTestGridProjects

Fluent builder constructing a request to ListTestGridProjects.

ListTestGridSessionActions

Fluent builder constructing a request to ListTestGridSessionActions.

ListTestGridSessionArtifacts

Fluent builder constructing a request to ListTestGridSessionArtifacts.

ListTestGridSessions

Fluent builder constructing a request to ListTestGridSessions.

ListTests

Fluent builder constructing a request to ListTests.

ListUniqueProblems

Fluent builder constructing a request to ListUniqueProblems.

ListUploads

Fluent builder constructing a request to ListUploads.

ListVPCEConfigurations

Fluent builder constructing a request to ListVPCEConfigurations.

PurchaseOffering

Fluent builder constructing a request to PurchaseOffering.

RenewOffering

Fluent builder constructing a request to RenewOffering.

ScheduleRun

Fluent builder constructing a request to ScheduleRun.

StopJob

Fluent builder constructing a request to StopJob.

StopRemoteAccessSession

Fluent builder constructing a request to StopRemoteAccessSession.

StopRun

Fluent builder constructing a request to StopRun.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateDeviceInstance

Fluent builder constructing a request to UpdateDeviceInstance.

UpdateDevicePool

Fluent builder constructing a request to UpdateDevicePool.

UpdateInstanceProfile

Fluent builder constructing a request to UpdateInstanceProfile.

UpdateNetworkProfile

Fluent builder constructing a request to UpdateNetworkProfile.

UpdateProject

Fluent builder constructing a request to UpdateProject.

UpdateTestGridProject

Fluent builder constructing a request to UpdateTestGridProject.

UpdateUpload

Fluent builder constructing a request to UpdateUpload.

UpdateVPCEConfiguration

Fluent builder constructing a request to UpdateVPCEConfiguration.