Struct aws_sdk_devicefarm::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for DeviceFarm_20150623.

This client allows ergonomic access to a DeviceFarm_20150623-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_device_pool(&self) -> CreateDevicePool<C, M, R>

pub fn create_instance_profile(&self) -> CreateInstanceProfile<C, M, R>

pub fn create_network_profile(&self) -> CreateNetworkProfile<C, M, R>

pub fn create_project(&self) -> CreateProject<C, M, R>

pub fn create_remote_access_session(&self) -> CreateRemoteAccessSession<C, M, R>

pub fn create_test_grid_project(&self) -> CreateTestGridProject<C, M, R>

pub fn create_test_grid_url(&self) -> CreateTestGridUrl<C, M, R>

pub fn create_upload(&self) -> CreateUpload<C, M, R>

pub fn create_vpce_configuration(&self) -> CreateVPCEConfiguration<C, M, R>

pub fn delete_device_pool(&self) -> DeleteDevicePool<C, M, R>

pub fn delete_instance_profile(&self) -> DeleteInstanceProfile<C, M, R>

pub fn delete_network_profile(&self) -> DeleteNetworkProfile<C, M, R>

pub fn delete_project(&self) -> DeleteProject<C, M, R>

pub fn delete_remote_access_session(&self) -> DeleteRemoteAccessSession<C, M, R>

pub fn delete_run(&self) -> DeleteRun<C, M, R>

pub fn delete_test_grid_project(&self) -> DeleteTestGridProject<C, M, R>

pub fn delete_upload(&self) -> DeleteUpload<C, M, R>

pub fn delete_vpce_configuration(&self) -> DeleteVPCEConfiguration<C, M, R>

pub fn get_account_settings(&self) -> GetAccountSettings<C, M, R>

pub fn get_device(&self) -> GetDevice<C, M, R>

pub fn get_device_instance(&self) -> GetDeviceInstance<C, M, R>

pub fn get_device_pool(&self) -> GetDevicePool<C, M, R>

pub fn get_device_pool_compatibility(
&self
) -> GetDevicePoolCompatibility<C, M, R>

pub fn get_instance_profile(&self) -> GetInstanceProfile<C, M, R>

pub fn get_job(&self) -> GetJob<C, M, R>

pub fn get_network_profile(&self) -> GetNetworkProfile<C, M, R>

pub fn get_offering_status(&self) -> GetOfferingStatus<C, M, R>

pub fn get_project(&self) -> GetProject<C, M, R>

pub fn get_remote_access_session(&self) -> GetRemoteAccessSession<C, M, R>

pub fn get_run(&self) -> GetRun<C, M, R>

pub fn get_suite(&self) -> GetSuite<C, M, R>

pub fn get_test(&self) -> GetTest<C, M, R>

pub fn get_test_grid_project(&self) -> GetTestGridProject<C, M, R>

pub fn get_test_grid_session(&self) -> GetTestGridSession<C, M, R>

pub fn get_upload(&self) -> GetUpload<C, M, R>

pub fn get_vpce_configuration(&self) -> GetVPCEConfiguration<C, M, R>

pub fn install_to_remote_access_session(
&self
) -> InstallToRemoteAccessSession<C, M, R>

pub fn list_artifacts(&self) -> ListArtifacts<C, M, R>

pub fn list_device_instances(&self) -> ListDeviceInstances<C, M, R>

pub fn list_device_pools(&self) -> ListDevicePools<C, M, R>

pub fn list_devices(&self) -> ListDevices<C, M, R>

pub fn list_instance_profiles(&self) -> ListInstanceProfiles<C, M, R>

pub fn list_jobs(&self) -> ListJobs<C, M, R>

pub fn list_network_profiles(&self) -> ListNetworkProfiles<C, M, R>

pub fn list_offering_promotions(&self) -> ListOfferingPromotions<C, M, R>

pub fn list_offerings(&self) -> ListOfferings<C, M, R>

pub fn list_offering_transactions(&self) -> ListOfferingTransactions<C, M, R>

pub fn list_projects(&self) -> ListProjects<C, M, R>

pub fn list_remote_access_sessions(&self) -> ListRemoteAccessSessions<C, M, R>

pub fn list_runs(&self) -> ListRuns<C, M, R>

pub fn list_samples(&self) -> ListSamples<C, M, R>

pub fn list_suites(&self) -> ListSuites<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_test_grid_projects(&self) -> ListTestGridProjects<C, M, R>

pub fn list_test_grid_session_actions(
&self
) -> ListTestGridSessionActions<C, M, R>

pub fn list_test_grid_session_artifacts(
&self
) -> ListTestGridSessionArtifacts<C, M, R>

pub fn list_test_grid_sessions(&self) -> ListTestGridSessions<C, M, R>

pub fn list_tests(&self) -> ListTests<C, M, R>

pub fn list_unique_problems(&self) -> ListUniqueProblems<C, M, R>

pub fn list_uploads(&self) -> ListUploads<C, M, R>

pub fn list_vpce_configurations(&self) -> ListVPCEConfigurations<C, M, R>

pub fn purchase_offering(&self) -> PurchaseOffering<C, M, R>

pub fn renew_offering(&self) -> RenewOffering<C, M, R>

pub fn schedule_run(&self) -> ScheduleRun<C, M, R>

pub fn stop_job(&self) -> StopJob<C, M, R>

pub fn stop_remote_access_session(&self) -> StopRemoteAccessSession<C, M, R>

pub fn stop_run(&self) -> StopRun<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_device_instance(&self) -> UpdateDeviceInstance<C, M, R>

pub fn update_device_pool(&self) -> UpdateDevicePool<C, M, R>

pub fn update_instance_profile(&self) -> UpdateInstanceProfile<C, M, R>

pub fn update_network_profile(&self) -> UpdateNetworkProfile<C, M, R>

pub fn update_project(&self) -> UpdateProject<C, M, R>

pub fn update_test_grid_project(&self) -> UpdateTestGridProject<C, M, R>

pub fn update_upload(&self) -> UpdateUpload<C, M, R>

pub fn update_vpce_configuration(&self) -> UpdateVPCEConfiguration<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self