Module aws_sdk_devicefarm::error

Modules

argument_exception

See ArgumentException

cannot_delete_exception

See CannotDeleteException

idempotency_exception

See IdempotencyException

internal_service_exception

See InternalServiceException

invalid_operation_exception

See InvalidOperationException

limit_exceeded_exception

See LimitExceededException

not_eligible_exception

See NotEligibleException

not_found_exception

See NotFoundException

service_account_exception

See ServiceAccountException

tag_operation_exception

See TagOperationException

tag_policy_exception

See TagPolicyException

too_many_tags_exception

See TooManyTagsException

Structs

ArgumentException

An invalid argument was specified.

CannotDeleteException

The requested object could not be deleted.

CreateDevicePoolError

CreateInstanceProfileError

CreateNetworkProfileError

CreateProjectError

CreateRemoteAccessSessionError

CreateTestGridProjectError

CreateTestGridUrlError

CreateUploadError

CreateVPCEConfigurationError

DeleteDevicePoolError

DeleteInstanceProfileError

DeleteNetworkProfileError

DeleteProjectError

DeleteRemoteAccessSessionError

DeleteRunError

DeleteTestGridProjectError

DeleteUploadError

DeleteVPCEConfigurationError

GetAccountSettingsError

GetDeviceError

GetDeviceInstanceError

GetDevicePoolCompatibilityError

GetDevicePoolError

GetInstanceProfileError

GetJobError

GetNetworkProfileError

GetOfferingStatusError

GetProjectError

GetRemoteAccessSessionError

GetRunError

GetSuiteError

GetTestError

GetTestGridProjectError

GetTestGridSessionError

GetUploadError

GetVPCEConfigurationError

IdempotencyException

An entity with the same name already exists.

InstallToRemoteAccessSessionError

InternalServiceException

An internal exception was raised in the service. Contact aws-devicefarm-support@amazon.com if you see this

InvalidOperationException

There was an error with the update request, or you do not have sufficient permissions

LimitExceededException

A limit was exceeded.

ListArtifactsError

ListDeviceInstancesError

ListDevicePoolsError

ListDevicesError

ListInstanceProfilesError

ListJobsError

ListNetworkProfilesError

ListOfferingPromotionsError

ListOfferingTransactionsError

ListOfferingsError

ListProjectsError

ListRemoteAccessSessionsError

ListRunsError

ListSamplesError

ListSuitesError

ListTagsForResourceError

ListTestGridProjectsError

ListTestGridSessionActionsError

ListTestGridSessionArtifactsError

ListTestGridSessionsError

ListTestsError

ListUniqueProblemsError

ListUploadsError

ListVPCEConfigurationsError

NotEligibleException

Exception gets thrown when a user is not eligible to perform the specified

NotFoundException

The specified entity was not found.

PurchaseOfferingError

RenewOfferingError

ScheduleRunError

ServiceAccountException

There was a problem with the service account.

StopJobError

StopRemoteAccessSessionError

StopRunError

TagOperationException

The operation was not successful. Try again.

TagPolicyException

The request doesn't comply with the AWS Identity and Access Management (IAM) tag

TagResourceError

TooManyTagsException

The list of tags on the repository is over the limit. The maximum number of tags that

UntagResourceError

UpdateDeviceInstanceError

UpdateDevicePoolError

UpdateInstanceProfileError

UpdateNetworkProfileError

UpdateProjectError

UpdateTestGridProjectError

UpdateUploadError

UpdateVPCEConfigurationError

Enums

CreateDevicePoolErrorKind

CreateInstanceProfileErrorKind

CreateNetworkProfileErrorKind

CreateProjectErrorKind

CreateRemoteAccessSessionErrorKind

CreateTestGridProjectErrorKind

CreateTestGridUrlErrorKind

CreateUploadErrorKind

CreateVPCEConfigurationErrorKind

DeleteDevicePoolErrorKind

DeleteInstanceProfileErrorKind

DeleteNetworkProfileErrorKind

DeleteProjectErrorKind

DeleteRemoteAccessSessionErrorKind

DeleteRunErrorKind

DeleteTestGridProjectErrorKind

DeleteUploadErrorKind

DeleteVPCEConfigurationErrorKind

GetAccountSettingsErrorKind

GetDeviceErrorKind

GetDeviceInstanceErrorKind

GetDevicePoolCompatibilityErrorKind

GetDevicePoolErrorKind

GetInstanceProfileErrorKind

GetJobErrorKind

GetNetworkProfileErrorKind

GetOfferingStatusErrorKind

GetProjectErrorKind

GetRemoteAccessSessionErrorKind

GetRunErrorKind

GetSuiteErrorKind

GetTestErrorKind

GetTestGridProjectErrorKind

GetTestGridSessionErrorKind

GetUploadErrorKind

GetVPCEConfigurationErrorKind

InstallToRemoteAccessSessionErrorKind

ListArtifactsErrorKind

ListDeviceInstancesErrorKind

ListDevicePoolsErrorKind

ListDevicesErrorKind

ListInstanceProfilesErrorKind

ListJobsErrorKind

ListNetworkProfilesErrorKind

ListOfferingPromotionsErrorKind

ListOfferingTransactionsErrorKind

ListOfferingsErrorKind

ListProjectsErrorKind

ListRemoteAccessSessionsErrorKind

ListRunsErrorKind

ListSamplesErrorKind

ListSuitesErrorKind

ListTagsForResourceErrorKind

ListTestGridProjectsErrorKind

ListTestGridSessionActionsErrorKind

ListTestGridSessionArtifactsErrorKind

ListTestGridSessionsErrorKind

ListTestsErrorKind

ListUniqueProblemsErrorKind

ListUploadsErrorKind

ListVPCEConfigurationsErrorKind

PurchaseOfferingErrorKind

RenewOfferingErrorKind

ScheduleRunErrorKind

StopJobErrorKind

StopRemoteAccessSessionErrorKind

StopRunErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDeviceInstanceErrorKind

UpdateDevicePoolErrorKind

UpdateInstanceProfileErrorKind

UpdateNetworkProfileErrorKind

UpdateProjectErrorKind

UpdateTestGridProjectErrorKind

UpdateUploadErrorKind

UpdateVPCEConfigurationErrorKind