Module aws_sdk_devicefarm::input

Modules

create_device_pool_input

See CreateDevicePoolInput

create_instance_profile_input

See CreateInstanceProfileInput

create_network_profile_input

See CreateNetworkProfileInput

create_project_input

See CreateProjectInput

create_remote_access_session_input

See CreateRemoteAccessSessionInput

create_test_grid_project_input

See CreateTestGridProjectInput

create_test_grid_url_input

See CreateTestGridUrlInput

create_upload_input

See CreateUploadInput

create_vpce_configuration_input

See CreateVpceConfigurationInput

delete_device_pool_input

See DeleteDevicePoolInput

delete_instance_profile_input

See DeleteInstanceProfileInput

delete_network_profile_input

See DeleteNetworkProfileInput

delete_project_input

See DeleteProjectInput

delete_remote_access_session_input

See DeleteRemoteAccessSessionInput

delete_run_input

See DeleteRunInput

delete_test_grid_project_input

See DeleteTestGridProjectInput

delete_upload_input

See DeleteUploadInput

delete_vpce_configuration_input

See DeleteVpceConfigurationInput

get_account_settings_input

See GetAccountSettingsInput

get_device_input

See GetDeviceInput

get_device_instance_input

See GetDeviceInstanceInput

get_device_pool_compatibility_input

See GetDevicePoolCompatibilityInput

get_device_pool_input

See GetDevicePoolInput

get_instance_profile_input

See GetInstanceProfileInput

get_job_input

See GetJobInput

get_network_profile_input

See GetNetworkProfileInput

get_offering_status_input

See GetOfferingStatusInput

get_project_input

See GetProjectInput

get_remote_access_session_input

See GetRemoteAccessSessionInput

get_run_input

See GetRunInput

get_suite_input

See GetSuiteInput

get_test_grid_project_input

See GetTestGridProjectInput

get_test_grid_session_input

See GetTestGridSessionInput

get_test_input

See GetTestInput

get_upload_input

See GetUploadInput

get_vpce_configuration_input

See GetVpceConfigurationInput

install_to_remote_access_session_input

See InstallToRemoteAccessSessionInput

list_artifacts_input

See ListArtifactsInput

list_device_instances_input

See ListDeviceInstancesInput

list_device_pools_input

See ListDevicePoolsInput

list_devices_input

See ListDevicesInput

list_instance_profiles_input

See ListInstanceProfilesInput

list_jobs_input

See ListJobsInput

list_network_profiles_input

See ListNetworkProfilesInput

list_offering_promotions_input

See ListOfferingPromotionsInput

list_offering_transactions_input

See ListOfferingTransactionsInput

list_offerings_input

See ListOfferingsInput

list_projects_input

See ListProjectsInput

list_remote_access_sessions_input

See ListRemoteAccessSessionsInput

list_runs_input

See ListRunsInput

list_samples_input

See ListSamplesInput

list_suites_input

See ListSuitesInput

list_tags_for_resource_input

See ListTagsForResourceInput

list_test_grid_projects_input

See ListTestGridProjectsInput

list_test_grid_session_actions_input

See ListTestGridSessionActionsInput

list_test_grid_session_artifacts_input

See ListTestGridSessionArtifactsInput

list_test_grid_sessions_input

See ListTestGridSessionsInput

list_tests_input

See ListTestsInput

list_unique_problems_input

See ListUniqueProblemsInput

list_uploads_input

See ListUploadsInput

list_vpce_configurations_input

See ListVpceConfigurationsInput

purchase_offering_input

See PurchaseOfferingInput

renew_offering_input

See RenewOfferingInput

schedule_run_input

See ScheduleRunInput

stop_job_input

See StopJobInput

stop_remote_access_session_input

See StopRemoteAccessSessionInput

stop_run_input

See StopRunInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_device_instance_input

See UpdateDeviceInstanceInput

update_device_pool_input

See UpdateDevicePoolInput

update_instance_profile_input

See UpdateInstanceProfileInput

update_network_profile_input

See UpdateNetworkProfileInput

update_project_input

See UpdateProjectInput

update_test_grid_project_input

See UpdateTestGridProjectInput

update_upload_input

See UpdateUploadInput

update_vpce_configuration_input

See UpdateVpceConfigurationInput

Structs

CreateDevicePoolInput

Represents a request to the create device pool operation.

CreateInstanceProfileInput

CreateNetworkProfileInput

CreateProjectInput

Represents a request to the create project operation.

CreateRemoteAccessSessionInput

Creates and submits a request to start a remote access session.

CreateTestGridProjectInput

CreateTestGridUrlInput

CreateUploadInput

Represents a request to the create upload operation.

CreateVpceConfigurationInput

DeleteDevicePoolInput

Represents a request to the delete device pool operation.

DeleteInstanceProfileInput

DeleteNetworkProfileInput

DeleteProjectInput

Represents a request to the delete project operation.

DeleteRemoteAccessSessionInput

Represents the request to delete the specified remote access session.

DeleteRunInput

Represents a request to the delete run operation.

DeleteTestGridProjectInput

DeleteUploadInput

Represents a request to the delete upload operation.

DeleteVpceConfigurationInput

GetAccountSettingsInput

Represents the request sent to retrieve the account settings.

GetDeviceInput

Represents a request to the get device request.

GetDeviceInstanceInput

GetDevicePoolCompatibilityInput

Represents a request to the get device pool compatibility operation.

GetDevicePoolInput

Represents a request to the get device pool operation.

GetInstanceProfileInput

GetJobInput

Represents a request to the get job operation.

GetNetworkProfileInput

GetOfferingStatusInput

Represents the request to retrieve the offering status for the specified customer

GetProjectInput

Represents a request to the get project operation.

GetRemoteAccessSessionInput

Represents the request to get information about the specified remote access

GetRunInput

Represents a request to the get run operation.

GetSuiteInput

Represents a request to the get suite operation.

GetTestGridProjectInput

GetTestGridSessionInput

GetTestInput

Represents a request to the get test operation.

GetUploadInput

Represents a request to the get upload operation.

GetVpceConfigurationInput

InstallToRemoteAccessSessionInput

Represents the request to install an Android application (in .apk format) or an iOS

ListArtifactsInput

Represents a request to the list artifacts operation.

ListDeviceInstancesInput

ListDevicePoolsInput

Represents the result of a list device pools request.

ListDevicesInput

Represents the result of a list devices request.

ListInstanceProfilesInput

ListJobsInput

Represents a request to the list jobs operation.

ListNetworkProfilesInput

ListOfferingPromotionsInput

ListOfferingTransactionsInput

Represents the request to list the offering transaction history.

ListOfferingsInput

Represents the request to list all offerings.

ListProjectsInput

Represents a request to the list projects operation.

ListRemoteAccessSessionsInput

Represents the request to return information about the remote access

ListRunsInput

Represents a request to the list runs operation.

ListSamplesInput

Represents a request to the list samples operation.

ListSuitesInput

Represents a request to the list suites operation.

ListTagsForResourceInput

ListTestGridProjectsInput

ListTestGridSessionActionsInput

ListTestGridSessionArtifactsInput

ListTestGridSessionsInput

ListTestsInput

Represents a request to the list tests operation.

ListUniqueProblemsInput

Represents a request to the list unique problems operation.

ListUploadsInput

Represents a request to the list uploads operation.

ListVpceConfigurationsInput

PurchaseOfferingInput

Represents a request for a purchase offering.

RenewOfferingInput

A request that represents an offering renewal.

ScheduleRunInput

Represents a request to the schedule run operation.

StopJobInput

StopRemoteAccessSessionInput

Represents the request to stop the remote access session.

StopRunInput

Represents the request to stop a specific run.

TagResourceInput

UntagResourceInput

UpdateDeviceInstanceInput

UpdateDevicePoolInput

Represents a request to the update device pool operation.

UpdateInstanceProfileInput

UpdateNetworkProfileInput

UpdateProjectInput

Represents a request to the update project operation.

UpdateTestGridProjectInput

UpdateUploadInput

UpdateVpceConfigurationInput