Module aws_sdk_devicefarm::model

Modules

account_settings

See AccountSettings

artifact

See Artifact

counters

See Counters

cpu

See Cpu

create_remote_access_session_configuration

See CreateRemoteAccessSessionConfiguration

customer_artifact_paths

See CustomerArtifactPaths

device

See Device

device_filter

See DeviceFilter

device_instance

See DeviceInstance

device_minutes

See DeviceMinutes

device_pool

See DevicePool

device_pool_compatibility_result

See DevicePoolCompatibilityResult

device_selection_configuration

See DeviceSelectionConfiguration

device_selection_result

See DeviceSelectionResult

execution_configuration

See ExecutionConfiguration

incompatibility_message

See IncompatibilityMessage

instance_profile

See InstanceProfile

job

See Job

location

See Location

monetary_amount

See MonetaryAmount

network_profile

See NetworkProfile

offering

See Offering

offering_promotion

See OfferingPromotion

offering_status

See OfferingStatus

offering_transaction

See OfferingTransaction

problem

See Problem

problem_detail

See ProblemDetail

project

See Project

radios

See Radios

recurring_charge

See RecurringCharge

remote_access_session

See RemoteAccessSession

resolution

See Resolution

rule

See Rule

run

See Run

sample

See Sample

schedule_run_configuration

See ScheduleRunConfiguration

schedule_run_test

See ScheduleRunTest

suite

See Suite

tag

See Tag

test

See Test

test_grid_project

See TestGridProject

test_grid_session

See TestGridSession

test_grid_session_action

See TestGridSessionAction

test_grid_session_artifact

See TestGridSessionArtifact

test_grid_vpc_config

See TestGridVpcConfig

trial_minutes

See TrialMinutes

unique_problem

See UniqueProblem

upload

See Upload

vpce_configuration

See VpceConfiguration

Structs

AccountSettings

A container for account-level settings in AWS Device Farm.

Artifact

Represents the output of a test. Examples of artifacts include logs and

Counters

Represents entity counters.

Cpu

Represents the amount of CPU that an app is using on a physical device. Does not represent system-wide

CreateRemoteAccessSessionConfiguration

Configuration settings for a remote access session, including billing

CustomerArtifactPaths

A JSON object that specifies the paths where the artifacts generated by the customer's tests, on the

Device

Represents a device type that an app is tested against.

DeviceFilter

Represents a device filter used to select a set of devices to be included in a test

DeviceInstance

Represents the device instance.

DeviceMinutes

Represents the total (metered or unmetered) minutes used by the resource to run

DevicePool

Represents a collection of device types.

DevicePoolCompatibilityResult

Represents a device pool compatibility result.

DeviceSelectionConfiguration

Represents the device filters used in a test run and the maximum number of devices to be included in the

DeviceSelectionResult

Contains the run results requested by the device selection configuration and how many devices were

ExecutionConfiguration

Represents configuration information about a test run, such as the execution

IncompatibilityMessage

Represents information about incompatibility.

InstanceProfile

Represents the instance profile.

Job

Represents a device.

Location

Represents a latitude and longitude pair, expressed in geographic coordinate system degrees (for

MonetaryAmount

A number that represents the monetary amount for an offering or transaction.

NetworkProfile

An array of settings that describes characteristics of a network profile.

Offering

Represents the metadata of a device offering.

OfferingPromotion

Represents information about an offering promotion.

OfferingStatus

The status of the offering.

OfferingTransaction

Represents the metadata of an offering transaction.

Problem

Represents a specific warning or failure.

ProblemDetail

Information about a problem detail.

Project

Represents an operating-system neutral workspace for running and managing

Radios

Represents the set of radios and their states on a device. Examples of radios

RecurringCharge

Specifies whether charges for devices are recurring.

RemoteAccessSession

Represents information about the remote access session.

Resolution

Represents the screen resolution of a device in height and width, expressed in

Rule

Represents a condition for a device pool.

Run

Represents a test run on a set of devices with a given app package, test parameters, and so

Sample

Represents a sample of performance data.

ScheduleRunConfiguration

Represents the settings for a run. Includes things like location, radio states,

ScheduleRunTest

Represents test settings. This data structure is passed in as the test parameter to ScheduleRun. For an

Suite

Represents a collection of one or more tests.

Tag

The metadata that you apply to a resource to help you categorize and organize it. Each tag consists of a

Test

Represents a condition that is evaluated.

TestGridProject

A Selenium testing project. Projects are used to collect and collate sessions.

TestGridSession

A TestGridSession is a single instance of a browser launched from the URL provided by a

TestGridSessionAction

An action taken by a TestGridSession browser instance.

TestGridSessionArtifact

Artifacts are video and other files that are produced in the process of running a browser in an automated

TestGridVpcConfig

The VPC security groups and subnets that are attached to a project.

TrialMinutes

Represents information about free trial device minutes for an AWS

UniqueProblem

A collection of one or more problems, grouped by their result.

Upload

An app or a set of one or more tests to upload or that have been

VpceConfiguration

Represents an Amazon Virtual Private Cloud (VPC) endpoint configuration.

Enums

ArtifactCategory

ArtifactType

NOTE: ArtifactType::Unknown has been renamed to ::UnknownValue.

BillingMethod

CurrencyCode

DeviceAttribute

DeviceAvailability

DeviceFilterAttribute

DeviceFormFactor

DevicePlatform

DevicePoolType

ExecutionResult

ExecutionResultCode

ExecutionStatus

InstanceStatus

InteractionMode

NetworkProfileType

OfferingTransactionType

OfferingType

RecurringChargeFrequency

RuleOperator

SampleType

TestGridSessionArtifactCategory

TestGridSessionArtifactType

NOTE: TestGridSessionArtifactType::Unknown has been renamed to ::UnknownValue.

TestGridSessionStatus

TestType

UploadCategory

UploadStatus

UploadType