Struct aws_sdk_devicefarm::model::device::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Device

Implementations

impl Builder

pub fn arn(self, input: impl Into<String>) -> Self

The device's ARN.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The device's display name.

pub fn set_name(self, input: Option<String>) -> Self

pub fn manufacturer(self, input: impl Into<String>) -> Self

The device's manufacturer name.

pub fn set_manufacturer(self, input: Option<String>) -> Self

pub fn model(self, input: impl Into<String>) -> Self

The device's model name.

pub fn set_model(self, input: Option<String>) -> Self

pub fn model_id(self, input: impl Into<String>) -> Self

The device's model ID.

pub fn set_model_id(self, input: Option<String>) -> Self

pub fn form_factor(self, input: DeviceFormFactor) -> Self

The device's form factor.

Allowed values include:

PHONE

TABLET

pub fn set_form_factor(self, input: Option<DeviceFormFactor>) -> Self

pub fn platform(self, input: DevicePlatform) -> Self

The device's platform.

Allowed values include:

ANDROID

IOS

pub fn set_platform(self, input: Option<DevicePlatform>) -> Self

pub fn os(self, input: impl Into<String>) -> Self

The device's operating system type.

pub fn set_os(self, input: Option<String>) -> Self

pub fn cpu(self, input: Cpu) -> Self

Information about the device's CPU.

pub fn set_cpu(self, input: Option<Cpu>) -> Self

pub fn resolution(self, input: Resolution) -> Self

The resolution of the device.

pub fn set_resolution(self, input: Option<Resolution>) -> Self

pub fn heap_size(self, input: i64) -> Self

The device's heap size, expressed in bytes.

pub fn set_heap_size(self, input: Option<i64>) -> Self

pub fn memory(self, input: i64) -> Self

The device's total memory size, expressed in bytes.

pub fn set_memory(self, input: Option<i64>) -> Self

pub fn image(self, input: impl Into<String>) -> Self

The device's image name.

pub fn set_image(self, input: Option<String>) -> Self

pub fn carrier(self, input: impl Into<String>) -> Self

The device's carrier.

pub fn set_carrier(self, input: Option<String>) -> Self

pub fn radio(self, input: impl Into<String>) -> Self

The device's radio.

pub fn set_radio(self, input: Option<String>) -> Self

pub fn remote_access_enabled(self, input: bool) -> Self

Specifies whether remote access has been enabled for the specified device.

pub fn set_remote_access_enabled(self, input: Option<bool>) -> Self

pub fn remote_debug_enabled(self, input: bool) -> Self

This flag is set to true if remote debugging is enabled for the device.

Remote debugging is no longer supported.

pub fn set_remote_debug_enabled(self, input: Option<bool>) -> Self

pub fn fleet_type(self, input: impl Into<String>) -> Self

The type of fleet to which this device belongs. Possible values are PRIVATE and PUBLIC.

pub fn set_fleet_type(self, input: Option<String>) -> Self

pub fn fleet_name(self, input: impl Into<String>) -> Self

The name of the fleet to which this device belongs.

pub fn set_fleet_name(self, input: Option<String>) -> Self

pub fn instances(self, input: impl Into<DeviceInstance>) -> Self

pub fn set_instances(self, input: Option<Vec<DeviceInstance>>) -> Self

pub fn availability(self, input: DeviceAvailability) -> Self

Indicates how likely a device is available for a test run. Currently available in the ListDevices and GetDevice API methods.

pub fn set_availability(self, input: Option<DeviceAvailability>) -> Self

pub fn build(self) -> Device

Consumes the builder and constructs a Device