Struct aws_sdk_devicefarm::model::Run

#[non_exhaustive]
pub struct Run {

    pub arn: Option<String>,
    pub name: Option<String>,
    pub type: Option<TestType>,
    pub platform: Option<DevicePlatform>,
    pub created: Option<Instant>,
    pub status: Option<ExecutionStatus>,
    pub result: Option<ExecutionResult>,
    pub started: Option<Instant>,
    pub stopped: Option<Instant>,
    pub counters: Option<Counters>,
    pub message: Option<String>,
    pub total_jobs: Option<i32>,
    pub completed_jobs: Option<i32>,
    pub billing_method: Option<BillingMethod>,
    pub device_minutes: Option<DeviceMinutes>,
    pub network_profile: Option<NetworkProfile>,
    pub parsing_result_url: Option<String>,
    pub result_code: Option<ExecutionResultCode>,
    pub seed: Option<i32>,
    pub app_upload: Option<String>,
    pub event_count: Option<i32>,
    pub job_timeout_minutes: Option<i32>,
    pub device_pool_arn: Option<String>,
    pub locale: Option<String>,
    pub radios: Option<Radios>,
    pub location: Option<Location>,
    pub customer_artifact_paths: Option<CustomerArtifactPaths>,
    pub web_url: Option<String>,
    pub skip_app_resign: Option<bool>,
    pub test_spec_arn: Option<String>,
    pub device_selection_result: Option<DeviceSelectionResult>,

}

Represents a test run on a set of devices with a given app package, test parameters, and so on.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

arn: Option<String>

The run's ARN.

name: Option<String>

The run's name.

type: Option<TestType>

The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

platform: Option<DevicePlatform>

The run's platform.

Allowed values include:

ANDROID

IOS

created: Option<Instant>

When the run was created.

status: Option<ExecutionStatus>

The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

result: Option<ExecutionResult>

The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

started: Option<Instant>

The run's start time.

stopped: Option<Instant>

The run's stop time.

counters: Option<Counters>

The run's result counters.

message: Option<String>

A message about the run's result.

total_jobs: Option<i32>

The total number of jobs for the run.

completed_jobs: Option<i32>

The total number of completed jobs.

billing_method: Option<BillingMethod>

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

device_minutes: Option<DeviceMinutes>

Represents the total (metered or unmetered) minutes used by the test run.

network_profile: Option<NetworkProfile>

The network profile being used for a test run.

parsing_result_url: Option<String>

Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

result_code: Option<ExecutionResultCode>

Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

seed: Option<i32>

For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

app_upload: Option<String>

An app to upload or that has been uploaded.

event_count: Option<i32>

For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

job_timeout_minutes: Option<i32>

The number of minutes the job executes before it times out.

device_pool_arn: Option<String>

The ARN of the device pool for the run.

locale: Option<String>

Information about the locale that is used for the run.

radios: Option<Radios>

Information about the radio states for the run.

location: Option<Location>

Information about the location that is used for the run.

customer_artifact_paths: Option<CustomerArtifactPaths>

Output CustomerArtifactPaths object for the test run.

web_url: Option<String>

The Device Farm console URL for the recording of the run.

skip_app_resign: Option<bool>

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

test_spec_arn: Option<String>

The ARN of the YAML-formatted test specification for the run.

device_selection_result: Option<DeviceSelectionResult>

The results of a device filter used to select the devices for a test run.

Implementations

impl Run

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Run