Module aws_sdk_devicefarm::operation

Structs

CreateDevicePool

Creates a device pool.

CreateInstanceProfile

Creates a profile that can be applied to one or more private fleet device

CreateNetworkProfile

Creates a network profile.

CreateProject

Creates a project.

CreateRemoteAccessSession

Specifies and starts a remote access session.

CreateTestGridProject

Creates a Selenium testing project. Projects are used to track TestGridSession

CreateTestGridUrl

Creates a signed, short-term URL that can be passed to a Selenium RemoteWebDriver

CreateUpload

Uploads an app or test scripts.

CreateVPCEConfiguration

Creates a configuration record in Device Farm for your Amazon Virtual Private Cloud

DeleteDevicePool

Deletes a device pool given the pool ARN. Does not allow deletion of curated pools

DeleteInstanceProfile

Deletes a profile that can be applied to one or more private device instances.

DeleteNetworkProfile

Deletes a network profile.

DeleteProject

Deletes an AWS Device Farm project, given the project ARN.

DeleteRemoteAccessSession

Deletes a completed remote access session and its results.

DeleteRun

Deletes the run, given the run ARN.

DeleteTestGridProject

Deletes a Selenium testing project and all content generated under it.

DeleteUpload

Deletes an upload given the upload ARN.

DeleteVPCEConfiguration

Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint.

GetAccountSettings

Returns the number of unmetered iOS or unmetered Android devices that have been purchased by the

GetDevice

Gets information about a unique device type.

GetDeviceInstance

Returns information about a device instance that belongs to a private device fleet.

GetDevicePool

Gets information about a device pool.

GetDevicePoolCompatibility

Gets information about compatibility with a device pool.

GetInstanceProfile

Returns information about the specified instance profile.

GetJob

Gets information about a job.

GetNetworkProfile

Returns information about a network profile.

GetOfferingStatus

Gets the current status and future status of all offerings purchased by an AWS account. The response

GetProject

Gets information about a project.

GetRemoteAccessSession

Returns a link to a currently running remote access session.

GetRun

Gets information about a run.

GetSuite

Gets information about a suite.

GetTest

Gets information about a test.

GetTestGridProject

Retrieves information about a Selenium testing project.

GetTestGridSession

A session is an instance of a browser created through a RemoteWebDriver with the URL from CreateTestGridUrlResult$url. You can use the following to look up sessions:

GetUpload

Gets information about an upload.

GetVPCEConfiguration

Returns information about the configuration settings for your Amazon Virtual Private

InstallToRemoteAccessSession

Installs an application to the device in a remote access session. For Android

ListArtifacts

Gets information about artifacts.

ListDeviceInstances

Returns information about the private device instances associated with one or more AWS

ListDevicePools

Gets information about device pools.

ListDevices

Gets information about unique device types.

ListInstanceProfiles

Returns information about all the instance profiles in an AWS account.

ListJobs

Gets information about jobs for a given test run.

ListNetworkProfiles

Returns the list of available network profiles.

ListOfferingPromotions

Returns a list of offering promotions. Each offering promotion record contains the ID and description

ListOfferingTransactions

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS

ListOfferings

Returns a list of products or offerings that the user can manage through the API. Each offering record

ListProjects

Gets information about projects.

ListRemoteAccessSessions

Returns a list of all currently running remote access sessions.

ListRuns

Gets information about runs, given an AWS Device Farm project ARN.

ListSamples

Gets information about samples, given an AWS Device Farm job ARN.

ListSuites

Gets information about test suites for a given job.

ListTagsForResource

List the tags for an AWS Device Farm resource.

ListTestGridProjects

Gets a list of all Selenium testing projects in your account.

ListTestGridSessionActions

Returns a list of the actions taken in a TestGridSession.

ListTestGridSessionArtifacts

Retrieves a list of artifacts created during the session.

ListTestGridSessions

Retrieves a list of sessions for a TestGridProject.

ListTests

Gets information about tests in a given test suite.

ListUniqueProblems

Gets information about unique problems, such as exceptions or crashes.

ListUploads

Gets information about uploads, given an AWS Device Farm project ARN.

ListVPCEConfigurations

Returns information about all Amazon Virtual Private Cloud (VPC) endpoint

PurchaseOffering

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased

RenewOffering

Explicitly sets the quantity of devices to renew for an offering, starting from the

ScheduleRun

Schedules a run.

StopJob

Initiates a stop request for the current job. AWS Device Farm immediately stops the job on the device

StopRemoteAccessSession

Ends a specified remote access session.

StopRun

Initiates a stop request for the current test run. AWS Device Farm immediately stops the run on devices

TagResource

Associates the specified tags to a resource with the specified resourceArn. If existing tags

UntagResource

Deletes the specified tags from a resource.

UpdateDeviceInstance

Updates information about a private device instance.

UpdateDevicePool

Modifies the name, description, and rules in a device pool given the attributes and

UpdateInstanceProfile

Updates information about an existing private device instance profile.

UpdateNetworkProfile

Updates the network profile.

UpdateProject

Modifies the specified project name, given the project ARN and a new

UpdateTestGridProject

Change details of a project.

UpdateUpload

Updates an uploaded test spec.

UpdateVPCEConfiguration

Updates information about an Amazon Virtual Private Cloud (VPC) endpoint configuration.