Module aws_sdk_devicefarm::output

Modules

create_device_pool_output

See CreateDevicePoolOutput

create_instance_profile_output

See CreateInstanceProfileOutput

create_network_profile_output

See CreateNetworkProfileOutput

create_project_output

See CreateProjectOutput

create_remote_access_session_output

See CreateRemoteAccessSessionOutput

create_test_grid_project_output

See CreateTestGridProjectOutput

create_test_grid_url_output

See CreateTestGridUrlOutput

create_upload_output

See CreateUploadOutput

create_vpce_configuration_output

See CreateVpceConfigurationOutput

delete_device_pool_output

See DeleteDevicePoolOutput

delete_instance_profile_output

See DeleteInstanceProfileOutput

delete_network_profile_output

See DeleteNetworkProfileOutput

delete_project_output

See DeleteProjectOutput

delete_remote_access_session_output

See DeleteRemoteAccessSessionOutput

delete_run_output

See DeleteRunOutput

delete_test_grid_project_output

See DeleteTestGridProjectOutput

delete_upload_output

See DeleteUploadOutput

delete_vpce_configuration_output

See DeleteVpceConfigurationOutput

get_account_settings_output

See GetAccountSettingsOutput

get_device_instance_output

See GetDeviceInstanceOutput

get_device_output

See GetDeviceOutput

get_device_pool_compatibility_output

See GetDevicePoolCompatibilityOutput

get_device_pool_output

See GetDevicePoolOutput

get_instance_profile_output

See GetInstanceProfileOutput

get_job_output

See GetJobOutput

get_network_profile_output

See GetNetworkProfileOutput

get_offering_status_output

See GetOfferingStatusOutput

get_project_output

See GetProjectOutput

get_remote_access_session_output

See GetRemoteAccessSessionOutput

get_run_output

See GetRunOutput

get_suite_output

See GetSuiteOutput

get_test_grid_project_output

See GetTestGridProjectOutput

get_test_grid_session_output

See GetTestGridSessionOutput

get_test_output

See GetTestOutput

get_upload_output

See GetUploadOutput

get_vpce_configuration_output

See GetVpceConfigurationOutput

install_to_remote_access_session_output

See InstallToRemoteAccessSessionOutput

list_artifacts_output

See ListArtifactsOutput

list_device_instances_output

See ListDeviceInstancesOutput

list_device_pools_output

See ListDevicePoolsOutput

list_devices_output

See ListDevicesOutput

list_instance_profiles_output

See ListInstanceProfilesOutput

list_jobs_output

See ListJobsOutput

list_network_profiles_output

See ListNetworkProfilesOutput

list_offering_promotions_output

See ListOfferingPromotionsOutput

list_offering_transactions_output

See ListOfferingTransactionsOutput

list_offerings_output

See ListOfferingsOutput

list_projects_output

See ListProjectsOutput

list_remote_access_sessions_output

See ListRemoteAccessSessionsOutput

list_runs_output

See ListRunsOutput

list_samples_output

See ListSamplesOutput

list_suites_output

See ListSuitesOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

list_test_grid_projects_output

See ListTestGridProjectsOutput

list_test_grid_session_actions_output

See ListTestGridSessionActionsOutput

list_test_grid_session_artifacts_output

See ListTestGridSessionArtifactsOutput

list_test_grid_sessions_output

See ListTestGridSessionsOutput

list_tests_output

See ListTestsOutput

list_unique_problems_output

See ListUniqueProblemsOutput

list_uploads_output

See ListUploadsOutput

list_vpce_configurations_output

See ListVpceConfigurationsOutput

purchase_offering_output

See PurchaseOfferingOutput

renew_offering_output

See RenewOfferingOutput

schedule_run_output

See ScheduleRunOutput

stop_job_output

See StopJobOutput

stop_remote_access_session_output

See StopRemoteAccessSessionOutput

stop_run_output

See StopRunOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_device_instance_output

See UpdateDeviceInstanceOutput

update_device_pool_output

See UpdateDevicePoolOutput

update_instance_profile_output

See UpdateInstanceProfileOutput

update_network_profile_output

See UpdateNetworkProfileOutput

update_project_output

See UpdateProjectOutput

update_test_grid_project_output

See UpdateTestGridProjectOutput

update_upload_output

See UpdateUploadOutput

update_vpce_configuration_output

See UpdateVpceConfigurationOutput

Structs

CreateDevicePoolOutput

Represents the result of a create device pool request.

CreateInstanceProfileOutput

CreateNetworkProfileOutput

CreateProjectOutput

Represents the result of a create project request.

CreateRemoteAccessSessionOutput

Represents the server response from a request to create a remote access

CreateTestGridProjectOutput

CreateTestGridUrlOutput

CreateUploadOutput

Represents the result of a create upload request.

CreateVpceConfigurationOutput

DeleteDevicePoolOutput

Represents the result of a delete device pool request.

DeleteInstanceProfileOutput

DeleteNetworkProfileOutput

DeleteProjectOutput

Represents the result of a delete project request.

DeleteRemoteAccessSessionOutput

The response from the server when a request is made to delete the remote access

DeleteRunOutput

Represents the result of a delete run request.

DeleteTestGridProjectOutput

DeleteUploadOutput

Represents the result of a delete upload request.

DeleteVpceConfigurationOutput

GetAccountSettingsOutput

Represents the account settings return values from the

GetDeviceInstanceOutput

GetDeviceOutput

Represents the result of a get device request.

GetDevicePoolCompatibilityOutput

Represents the result of describe device pool compatibility request.

GetDevicePoolOutput

Represents the result of a get device pool request.

GetInstanceProfileOutput

GetJobOutput

Represents the result of a get job request.

GetNetworkProfileOutput

GetOfferingStatusOutput

Returns the status result for a device offering.

GetProjectOutput

Represents the result of a get project request.

GetRemoteAccessSessionOutput

Represents the response from the server that lists detailed information about the

GetRunOutput

Represents the result of a get run request.

GetSuiteOutput

Represents the result of a get suite request.

GetTestGridProjectOutput

GetTestGridSessionOutput

GetTestOutput

Represents the result of a get test request.

GetUploadOutput

Represents the result of a get upload request.

GetVpceConfigurationOutput

InstallToRemoteAccessSessionOutput

Represents the response from the server after AWS Device Farm makes a request to

ListArtifactsOutput

Represents the result of a list artifacts operation.

ListDeviceInstancesOutput

ListDevicePoolsOutput

Represents the result of a list device pools request.

ListDevicesOutput

Represents the result of a list devices operation.

ListInstanceProfilesOutput

ListJobsOutput

Represents the result of a list jobs request.

ListNetworkProfilesOutput

ListOfferingPromotionsOutput

ListOfferingTransactionsOutput

Returns the transaction log of the specified offerings.

ListOfferingsOutput

Represents the return values of the list of offerings.

ListProjectsOutput

Represents the result of a list projects request.

ListRemoteAccessSessionsOutput

Represents the response from the server after AWS Device Farm makes a request to

ListRunsOutput

Represents the result of a list runs request.

ListSamplesOutput

Represents the result of a list samples request.

ListSuitesOutput

Represents the result of a list suites request.

ListTagsForResourceOutput

ListTestGridProjectsOutput

ListTestGridSessionActionsOutput

ListTestGridSessionArtifactsOutput

ListTestGridSessionsOutput

ListTestsOutput

Represents the result of a list tests request.

ListUniqueProblemsOutput

Represents the result of a list unique problems request.

ListUploadsOutput

Represents the result of a list uploads request.

ListVpceConfigurationsOutput

PurchaseOfferingOutput

The result of the purchase offering (for example, success or failure).

RenewOfferingOutput

The result of a renewal offering.

ScheduleRunOutput

Represents the result of a schedule run request.

StopJobOutput

StopRemoteAccessSessionOutput

Represents the response from the server that describes the remote access session

StopRunOutput

Represents the results of your stop run attempt.

TagResourceOutput

UntagResourceOutput

UpdateDeviceInstanceOutput

UpdateDevicePoolOutput

Represents the result of an update device pool request.

UpdateInstanceProfileOutput

UpdateNetworkProfileOutput

UpdateProjectOutput

Represents the result of an update project request.

UpdateTestGridProjectOutput

UpdateUploadOutput

UpdateVpceConfigurationOutput