Struct aws_sdk_devicefarm::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for DeviceFarm_20150623.

This client allows ergonomic access to a DeviceFarm_20150623-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_device_pool(&self) -> CreateDevicePool<C, M, R>

Constructs a fluent builder for the CreateDevicePool operation.

See CreateDevicePool for more information about the operation and its arguments.

pub fn create_instance_profile(&self) -> CreateInstanceProfile<C, M, R>

Constructs a fluent builder for the CreateInstanceProfile operation.

See CreateInstanceProfile for more information about the operation and its arguments.

pub fn create_network_profile(&self) -> CreateNetworkProfile<C, M, R>

Constructs a fluent builder for the CreateNetworkProfile operation.

See CreateNetworkProfile for more information about the operation and its arguments.

pub fn create_project(&self) -> CreateProject<C, M, R>

Constructs a fluent builder for the CreateProject operation.

See CreateProject for more information about the operation and its arguments.

pub fn create_remote_access_session(&self) -> CreateRemoteAccessSession<C, M, R>

Constructs a fluent builder for the CreateRemoteAccessSession operation.

See CreateRemoteAccessSession for more information about the operation and its arguments.

pub fn create_test_grid_project(&self) -> CreateTestGridProject<C, M, R>

Constructs a fluent builder for the CreateTestGridProject operation.

See CreateTestGridProject for more information about the operation and its arguments.

pub fn create_test_grid_url(&self) -> CreateTestGridUrl<C, M, R>

Constructs a fluent builder for the CreateTestGridUrl operation.

See CreateTestGridUrl for more information about the operation and its arguments.

pub fn create_upload(&self) -> CreateUpload<C, M, R>

Constructs a fluent builder for the CreateUpload operation.

See CreateUpload for more information about the operation and its arguments.

pub fn create_vpce_configuration(&self) -> CreateVPCEConfiguration<C, M, R>

Constructs a fluent builder for the CreateVPCEConfiguration operation.

See CreateVPCEConfiguration for more information about the operation and its arguments.

pub fn delete_device_pool(&self) -> DeleteDevicePool<C, M, R>

Constructs a fluent builder for the DeleteDevicePool operation.

See DeleteDevicePool for more information about the operation and its arguments.

pub fn delete_instance_profile(&self) -> DeleteInstanceProfile<C, M, R>

Constructs a fluent builder for the DeleteInstanceProfile operation.

See DeleteInstanceProfile for more information about the operation and its arguments.

pub fn delete_network_profile(&self) -> DeleteNetworkProfile<C, M, R>

Constructs a fluent builder for the DeleteNetworkProfile operation.

See DeleteNetworkProfile for more information about the operation and its arguments.

pub fn delete_project(&self) -> DeleteProject<C, M, R>

Constructs a fluent builder for the DeleteProject operation.

See DeleteProject for more information about the operation and its arguments.

pub fn delete_remote_access_session(&self) -> DeleteRemoteAccessSession<C, M, R>

Constructs a fluent builder for the DeleteRemoteAccessSession operation.

See DeleteRemoteAccessSession for more information about the operation and its arguments.

pub fn delete_run(&self) -> DeleteRun<C, M, R>

Constructs a fluent builder for the DeleteRun operation.

See DeleteRun for more information about the operation and its arguments.

pub fn delete_test_grid_project(&self) -> DeleteTestGridProject<C, M, R>

Constructs a fluent builder for the DeleteTestGridProject operation.

See DeleteTestGridProject for more information about the operation and its arguments.

pub fn delete_upload(&self) -> DeleteUpload<C, M, R>

Constructs a fluent builder for the DeleteUpload operation.

See DeleteUpload for more information about the operation and its arguments.

pub fn delete_vpce_configuration(&self) -> DeleteVPCEConfiguration<C, M, R>

Constructs a fluent builder for the DeleteVPCEConfiguration operation.

See DeleteVPCEConfiguration for more information about the operation and its arguments.

pub fn get_account_settings(&self) -> GetAccountSettings<C, M, R>

Constructs a fluent builder for the GetAccountSettings operation.

See GetAccountSettings for more information about the operation and its arguments.

pub fn get_device(&self) -> GetDevice<C, M, R>

Constructs a fluent builder for the GetDevice operation.

See GetDevice for more information about the operation and its arguments.

pub fn get_device_instance(&self) -> GetDeviceInstance<C, M, R>

Constructs a fluent builder for the GetDeviceInstance operation.

See GetDeviceInstance for more information about the operation and its arguments.

pub fn get_device_pool(&self) -> GetDevicePool<C, M, R>

Constructs a fluent builder for the GetDevicePool operation.

See GetDevicePool for more information about the operation and its arguments.

pub fn get_device_pool_compatibility(
&self
) -> GetDevicePoolCompatibility<C, M, R>

Constructs a fluent builder for the GetDevicePoolCompatibility operation.

See GetDevicePoolCompatibility for more information about the operation and its arguments.

pub fn get_instance_profile(&self) -> GetInstanceProfile<C, M, R>

Constructs a fluent builder for the GetInstanceProfile operation.

See GetInstanceProfile for more information about the operation and its arguments.

pub fn get_job(&self) -> GetJob<C, M, R>

Constructs a fluent builder for the GetJob operation.

See GetJob for more information about the operation and its arguments.

pub fn get_network_profile(&self) -> GetNetworkProfile<C, M, R>

Constructs a fluent builder for the GetNetworkProfile operation.

See GetNetworkProfile for more information about the operation and its arguments.

pub fn get_offering_status(&self) -> GetOfferingStatus<C, M, R>

Constructs a fluent builder for the GetOfferingStatus operation.

See GetOfferingStatus for more information about the operation and its arguments.

pub fn get_project(&self) -> GetProject<C, M, R>

Constructs a fluent builder for the GetProject operation.

See GetProject for more information about the operation and its arguments.

pub fn get_remote_access_session(&self) -> GetRemoteAccessSession<C, M, R>

Constructs a fluent builder for the GetRemoteAccessSession operation.

See GetRemoteAccessSession for more information about the operation and its arguments.

pub fn get_run(&self) -> GetRun<C, M, R>

Constructs a fluent builder for the GetRun operation.

See GetRun for more information about the operation and its arguments.

pub fn get_suite(&self) -> GetSuite<C, M, R>

Constructs a fluent builder for the GetSuite operation.

See GetSuite for more information about the operation and its arguments.

pub fn get_test(&self) -> GetTest<C, M, R>

Constructs a fluent builder for the GetTest operation.

See GetTest for more information about the operation and its arguments.

pub fn get_test_grid_project(&self) -> GetTestGridProject<C, M, R>

Constructs a fluent builder for the GetTestGridProject operation.

See GetTestGridProject for more information about the operation and its arguments.

pub fn get_test_grid_session(&self) -> GetTestGridSession<C, M, R>

Constructs a fluent builder for the GetTestGridSession operation.

See GetTestGridSession for more information about the operation and its arguments.

pub fn get_upload(&self) -> GetUpload<C, M, R>

Constructs a fluent builder for the GetUpload operation.

See GetUpload for more information about the operation and its arguments.

pub fn get_vpce_configuration(&self) -> GetVPCEConfiguration<C, M, R>

Constructs a fluent builder for the GetVPCEConfiguration operation.

See GetVPCEConfiguration for more information about the operation and its arguments.

pub fn install_to_remote_access_session(
&self
) -> InstallToRemoteAccessSession<C, M, R>

Constructs a fluent builder for the InstallToRemoteAccessSession operation.

See InstallToRemoteAccessSession for more information about the operation and its arguments.

pub fn list_artifacts(&self) -> ListArtifacts<C, M, R>

Constructs a fluent builder for the ListArtifacts operation.

See ListArtifacts for more information about the operation and its arguments.

pub fn list_device_instances(&self) -> ListDeviceInstances<C, M, R>

Constructs a fluent builder for the ListDeviceInstances operation.

See ListDeviceInstances for more information about the operation and its arguments.

pub fn list_device_pools(&self) -> ListDevicePools<C, M, R>

Constructs a fluent builder for the ListDevicePools operation.

See ListDevicePools for more information about the operation and its arguments.

pub fn list_devices(&self) -> ListDevices<C, M, R>

Constructs a fluent builder for the ListDevices operation.

See ListDevices for more information about the operation and its arguments.

pub fn list_instance_profiles(&self) -> ListInstanceProfiles<C, M, R>

Constructs a fluent builder for the ListInstanceProfiles operation.

See ListInstanceProfiles for more information about the operation and its arguments.

pub fn list_jobs(&self) -> ListJobs<C, M, R>

Constructs a fluent builder for the ListJobs operation.

See ListJobs for more information about the operation and its arguments.

pub fn list_network_profiles(&self) -> ListNetworkProfiles<C, M, R>

Constructs a fluent builder for the ListNetworkProfiles operation.

See ListNetworkProfiles for more information about the operation and its arguments.

pub fn list_offering_promotions(&self) -> ListOfferingPromotions<C, M, R>

Constructs a fluent builder for the ListOfferingPromotions operation.

See ListOfferingPromotions for more information about the operation and its arguments.

pub fn list_offerings(&self) -> ListOfferings<C, M, R>

Constructs a fluent builder for the ListOfferings operation.

See ListOfferings for more information about the operation and its arguments.

pub fn list_offering_transactions(&self) -> ListOfferingTransactions<C, M, R>

Constructs a fluent builder for the ListOfferingTransactions operation.

See ListOfferingTransactions for more information about the operation and its arguments.

pub fn list_projects(&self) -> ListProjects<C, M, R>

Constructs a fluent builder for the ListProjects operation.

See ListProjects for more information about the operation and its arguments.

pub fn list_remote_access_sessions(&self) -> ListRemoteAccessSessions<C, M, R>

Constructs a fluent builder for the ListRemoteAccessSessions operation.

See ListRemoteAccessSessions for more information about the operation and its arguments.

pub fn list_runs(&self) -> ListRuns<C, M, R>

Constructs a fluent builder for the ListRuns operation.

See ListRuns for more information about the operation and its arguments.

pub fn list_samples(&self) -> ListSamples<C, M, R>

Constructs a fluent builder for the ListSamples operation.

See ListSamples for more information about the operation and its arguments.

pub fn list_suites(&self) -> ListSuites<C, M, R>

Constructs a fluent builder for the ListSuites operation.

See ListSuites for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_test_grid_projects(&self) -> ListTestGridProjects<C, M, R>

Constructs a fluent builder for the ListTestGridProjects operation.

See ListTestGridProjects for more information about the operation and its arguments.

pub fn list_test_grid_session_actions(
&self
) -> ListTestGridSessionActions<C, M, R>

Constructs a fluent builder for the ListTestGridSessionActions operation.

See ListTestGridSessionActions for more information about the operation and its arguments.

pub fn list_test_grid_session_artifacts(
&self
) -> ListTestGridSessionArtifacts<C, M, R>

Constructs a fluent builder for the ListTestGridSessionArtifacts operation.

See ListTestGridSessionArtifacts for more information about the operation and its arguments.

pub fn list_test_grid_sessions(&self) -> ListTestGridSessions<C, M, R>

Constructs a fluent builder for the ListTestGridSessions operation.

See ListTestGridSessions for more information about the operation and its arguments.

pub fn list_tests(&self) -> ListTests<C, M, R>

Constructs a fluent builder for the ListTests operation.

See ListTests for more information about the operation and its arguments.

pub fn list_unique_problems(&self) -> ListUniqueProblems<C, M, R>

Constructs a fluent builder for the ListUniqueProblems operation.

See ListUniqueProblems for more information about the operation and its arguments.

pub fn list_uploads(&self) -> ListUploads<C, M, R>

Constructs a fluent builder for the ListUploads operation.

See ListUploads for more information about the operation and its arguments.

pub fn list_vpce_configurations(&self) -> ListVPCEConfigurations<C, M, R>

Constructs a fluent builder for the ListVPCEConfigurations operation.

See ListVPCEConfigurations for more information about the operation and its arguments.

pub fn purchase_offering(&self) -> PurchaseOffering<C, M, R>

Constructs a fluent builder for the PurchaseOffering operation.

See PurchaseOffering for more information about the operation and its arguments.

pub fn renew_offering(&self) -> RenewOffering<C, M, R>

Constructs a fluent builder for the RenewOffering operation.

See RenewOffering for more information about the operation and its arguments.

pub fn schedule_run(&self) -> ScheduleRun<C, M, R>

Constructs a fluent builder for the ScheduleRun operation.

See ScheduleRun for more information about the operation and its arguments.

pub fn stop_job(&self) -> StopJob<C, M, R>

Constructs a fluent builder for the StopJob operation.

See StopJob for more information about the operation and its arguments.

pub fn stop_remote_access_session(&self) -> StopRemoteAccessSession<C, M, R>

Constructs a fluent builder for the StopRemoteAccessSession operation.

See StopRemoteAccessSession for more information about the operation and its arguments.

pub fn stop_run(&self) -> StopRun<C, M, R>

Constructs a fluent builder for the StopRun operation.

See StopRun for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_device_instance(&self) -> UpdateDeviceInstance<C, M, R>

Constructs a fluent builder for the UpdateDeviceInstance operation.

See UpdateDeviceInstance for more information about the operation and its arguments.

pub fn update_device_pool(&self) -> UpdateDevicePool<C, M, R>

Constructs a fluent builder for the UpdateDevicePool operation.

See UpdateDevicePool for more information about the operation and its arguments.

pub fn update_instance_profile(&self) -> UpdateInstanceProfile<C, M, R>

Constructs a fluent builder for the UpdateInstanceProfile operation.

See UpdateInstanceProfile for more information about the operation and its arguments.

pub fn update_network_profile(&self) -> UpdateNetworkProfile<C, M, R>

Constructs a fluent builder for the UpdateNetworkProfile operation.

See UpdateNetworkProfile for more information about the operation and its arguments.

pub fn update_project(&self) -> UpdateProject<C, M, R>

Constructs a fluent builder for the UpdateProject operation.

See UpdateProject for more information about the operation and its arguments.

pub fn update_test_grid_project(&self) -> UpdateTestGridProject<C, M, R>

Constructs a fluent builder for the UpdateTestGridProject operation.

See UpdateTestGridProject for more information about the operation and its arguments.

pub fn update_upload(&self) -> UpdateUpload<C, M, R>

Constructs a fluent builder for the UpdateUpload operation.

See UpdateUpload for more information about the operation and its arguments.

pub fn update_vpce_configuration(&self) -> UpdateVPCEConfiguration<C, M, R>

Constructs a fluent builder for the UpdateVPCEConfiguration operation.

See UpdateVPCEConfiguration for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.