Struct aws_sdk_devicefarm::model::run::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Run

Implementations

impl Builder

pub fn arn(self, input: impl Into<String>) -> Self

The run's ARN.

pub fn set_arn(self, input: Option<String>) -> Self

The run's ARN.

pub fn name(self, input: impl Into<String>) -> Self

The run's name.

pub fn set_name(self, input: Option<String>) -> Self

The run's name.

pub fn type(self, input: TestType) -> Self

The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

pub fn set_type(self, input: Option<TestType>) -> Self

The run's type.

Must be one of the following values:

BUILTIN_FUZZ

BUILTIN_EXPLORER

For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT

APPIUM_JAVA_TESTNG

APPIUM_PYTHON

APPIUM_NODE

APPIUM_RUBY

APPIUM_WEB_JAVA_JUNIT

APPIUM_WEB_JAVA_TESTNG

APPIUM_WEB_PYTHON

APPIUM_WEB_NODE

APPIUM_WEB_RUBY

CALABASH

INSTRUMENTATION

UIAUTOMATION

UIAUTOMATOR

XCTEST

XCTEST_UI

pub fn platform(self, input: DevicePlatform) -> Self

The run's platform.

Allowed values include:

ANDROID

IOS

pub fn set_platform(self, input: Option<DevicePlatform>) -> Self

The run's platform.

Allowed values include:

ANDROID

IOS

pub fn created(self, input: Instant) -> Self

When the run was created.

pub fn set_created(self, input: Option<Instant>) -> Self

When the run was created.

pub fn status(self, input: ExecutionStatus) -> Self

The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

pub fn set_status(self, input: Option<ExecutionStatus>) -> Self

The run's status.

Allowed values include:

PENDING

PENDING_CONCURRENCY

PENDING_DEVICE

PROCESSING

SCHEDULING

PREPARING

RUNNING

COMPLETED

STOPPING

pub fn result(self, input: ExecutionResult) -> Self

The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

pub fn set_result(self, input: Option<ExecutionResult>) -> Self

The run's result.

Allowed values include:

PENDING

PASSED

WARNED

FAILED

SKIPPED

ERRORED

STOPPED

pub fn started(self, input: Instant) -> Self

The run's start time.

pub fn set_started(self, input: Option<Instant>) -> Self

The run's start time.

pub fn stopped(self, input: Instant) -> Self

The run's stop time.

pub fn set_stopped(self, input: Option<Instant>) -> Self

The run's stop time.

pub fn counters(self, input: Counters) -> Self

The run's result counters.

pub fn set_counters(self, input: Option<Counters>) -> Self

The run's result counters.

pub fn message(self, input: impl Into<String>) -> Self

A message about the run's result.

pub fn set_message(self, input: Option<String>) -> Self

A message about the run's result.

pub fn total_jobs(self, input: i32) -> Self

The total number of jobs for the run.

pub fn set_total_jobs(self, input: Option<i32>) -> Self

The total number of jobs for the run.

pub fn completed_jobs(self, input: i32) -> Self

The total number of completed jobs.

pub fn set_completed_jobs(self, input: Option<i32>) -> Self

The total number of completed jobs.

pub fn billing_method(self, input: BillingMethod) -> Self

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

pub fn set_billing_method(self, input: Option<BillingMethod>) -> Self

Specifies the billing method for a test run: metered or unmetered. If the parameter is not specified, the default value is metered.

If you have unmetered device slots, you must set this to unmetered to use them. Otherwise, the run is counted toward metered device minutes.

pub fn device_minutes(self, input: DeviceMinutes) -> Self

Represents the total (metered or unmetered) minutes used by the test run.

pub fn set_device_minutes(self, input: Option<DeviceMinutes>) -> Self

Represents the total (metered or unmetered) minutes used by the test run.

pub fn network_profile(self, input: NetworkProfile) -> Self

The network profile being used for a test run.

pub fn set_network_profile(self, input: Option<NetworkProfile>) -> Self

The network profile being used for a test run.

pub fn parsing_result_url(self, input: impl Into<String>) -> Self

Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

pub fn set_parsing_result_url(self, input: Option<String>) -> Self

Read-only URL for an object in an S3 bucket where you can get the parsing results of the test package. If the test package doesn't parse, the reason why it doesn't parse appears in the file that this URL points to.

pub fn result_code(self, input: ExecutionResultCode) -> Self

Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

pub fn set_result_code(self, input: Option<ExecutionResultCode>) -> Self

Supporting field for the result field. Set only if result is SKIPPED. PARSING_FAILED if the result is skipped because of test package parsing failure.

pub fn seed(self, input: i32) -> Self

For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

pub fn set_seed(self, input: Option<i32>) -> Self

For fuzz tests, this is a seed to use for randomizing the UI fuzz test. Using the same seed value between tests ensures identical event sequences.

pub fn app_upload(self, input: impl Into<String>) -> Self

An app to upload or that has been uploaded.

pub fn set_app_upload(self, input: Option<String>) -> Self

An app to upload or that has been uploaded.

pub fn event_count(self, input: i32) -> Self

For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

pub fn set_event_count(self, input: Option<i32>) -> Self

For fuzz tests, this is the number of events, between 1 and 10000, that the UI fuzz test should perform.

pub fn job_timeout_minutes(self, input: i32) -> Self

The number of minutes the job executes before it times out.

pub fn set_job_timeout_minutes(self, input: Option<i32>) -> Self

The number of minutes the job executes before it times out.

pub fn device_pool_arn(self, input: impl Into<String>) -> Self

The ARN of the device pool for the run.

pub fn set_device_pool_arn(self, input: Option<String>) -> Self

The ARN of the device pool for the run.

pub fn locale(self, input: impl Into<String>) -> Self

Information about the locale that is used for the run.

pub fn set_locale(self, input: Option<String>) -> Self

Information about the locale that is used for the run.

pub fn radios(self, input: Radios) -> Self

Information about the radio states for the run.

pub fn set_radios(self, input: Option<Radios>) -> Self

Information about the radio states for the run.

pub fn location(self, input: Location) -> Self

Information about the location that is used for the run.

pub fn set_location(self, input: Option<Location>) -> Self

Information about the location that is used for the run.

pub fn customer_artifact_paths(self, input: CustomerArtifactPaths) -> Self

Output CustomerArtifactPaths object for the test run.

pub fn set_customer_artifact_paths(
self,
input: Option<CustomerArtifactPaths>
) -> Self

Output CustomerArtifactPaths object for the test run.

pub fn web_url(self, input: impl Into<String>) -> Self

The Device Farm console URL for the recording of the run.

pub fn set_web_url(self, input: Option<String>) -> Self

The Device Farm console URL for the recording of the run.

pub fn skip_app_resign(self, input: bool) -> Self

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

pub fn set_skip_app_resign(self, input: Option<bool>) -> Self

When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again.

For more information about how Device Farm re-signs your apps, see Do you modify my app? in the AWS Device Farm FAQs.

pub fn test_spec_arn(self, input: impl Into<String>) -> Self

The ARN of the YAML-formatted test specification for the run.

pub fn set_test_spec_arn(self, input: Option<String>) -> Self

The ARN of the YAML-formatted test specification for the run.

pub fn device_selection_result(self, input: DeviceSelectionResult) -> Self

The results of a device filter used to select the devices for a test run.

pub fn set_device_selection_result(
self,
input: Option<DeviceSelectionResult>
) -> Self

The results of a device filter used to select the devices for a test run.

pub fn build(self) -> Run

Consumes the builder and constructs a Run